requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.1.0
Flask-Login==0.6.3
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.0
pytz==2024.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
twilio==9.6.2